begin moving to new video package
