+++ clients/transcode_provider.go
@@ -3,6 +3,7 @@ package clients
 import (
 	"context"
 	"fmt"
+	"github.com/livepeer/catalyst-api/video"
 	"net/url"
 )
 
@@ -15,8 +16,8 @@ type TranscodeJobArgs struct {
 	// Function that should be called every so often with the progress of the job.
 	ReportProgress func(completionRatio float64)
 	// Input File info used to by transcoder provider(s) to set transcode options
-	InputFileInfo InputVideo
-	Profiles      []EncodedProfile
+	InputFileInfo video.InputVideo
+	Profiles      []video.EncodedProfile
 
 	// Collect size of an asset
 	CollectSourceSize        func(size int64)
