+++ handlers/transcode.go
@@ -2,11 +2,11 @@ package handlers
 
 import (
 	"encoding/json"
+	"github.com/livepeer/catalyst-api/video"
 	"io"
 	"net/http"
 
 	"github.com/julienschmidt/httprouter"
-	"github.com/livepeer/catalyst-api/clients"
 	"github.com/livepeer/catalyst-api/errors"
 	"github.com/livepeer/catalyst-api/transcode"
 	"github.com/xeipuuv/gojsonschema"
@@ -38,7 +38,7 @@ func (d *CatalystAPIHandlersCollection) TranscodeSegment() httprouter.Handle {
 
 		// Note: This is no longer used pipeline stage
 		// TODO: Revisit later with better mistserver integration
-		_, _, err = transcode.RunTranscodeProcess(transcodeRequest, "", clients.InputVideo{})
+		_, _, err = transcode.RunTranscodeProcess(transcodeRequest, "", video.InputVideo{})
 		if err != nil {
 			errors.WriteHTTPInternalServerError(w, "Error running Transcode process", err)
 		}
