+++ handlers/upload.go
@@ -3,6 +3,7 @@ package handlers
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/livepeer/catalyst-api/video"
 	"io"
 	"mime"
 	"net/http"
@@ -12,7 +13,6 @@ import (
 	"time"
 
 	"github.com/julienschmidt/httprouter"
-	"github.com/livepeer/catalyst-api/clients"
 	"github.com/livepeer/catalyst-api/config"
 	"github.com/livepeer/catalyst-api/errors"
 	"github.com/livepeer/catalyst-api/log"
@@ -37,8 +37,8 @@ type UploadVODRequest struct {
 	AccessToken     string `json:"accessToken"`
 	TranscodeAPIUrl string `json:"transcodeAPIUrl"`
 	// Forwarded to transcoding stage:
-	Profiles         []clients.EncodedProfile `json:"profiles"`
-	PipelineStrategy pipeline.Strategy        `json:"pipeline_strategy"`
+	Profiles         []video.EncodedProfile `json:"profiles"`
+	PipelineStrategy pipeline.Strategy      `json:"pipeline_strategy"`
 }
 
 type UploadVODResponse struct {
