+++ transcode/transcode_test.go
@@ -3,6 +3,7 @@ package transcode
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/livepeer/catalyst-api/video"
 	"io"
 	"net/http"
 	"net/http/httptest"
@@ -34,7 +35,7 @@ type StubBroadcasterClient struct {
 	tr clients.TranscodeResult
 }
 
-func (c StubBroadcasterClient) TranscodeSegment(segment io.Reader, sequenceNumber int64, profiles []clients.EncodedProfile, durationMillis int64, manifestID string) (clients.TranscodeResult, error) {
+func (c StubBroadcasterClient) TranscodeSegment(segment io.Reader, sequenceNumber int64, profiles []video.EncodedProfile, durationMillis int64, manifestID string) (clients.TranscodeResult, error) {
 	return c.tr, nil
 }
 
@@ -86,7 +87,7 @@ func TestItCanTranscode(t *testing.T) {
 	}))
 	defer callbackServer.Close()
 
-	sourceVideoTrack := clients.VideoTrack{
+	sourceVideoTrack := video.VideoTrack{
 		Width:  2020,
 		Height: 2020,
 	}
@@ -118,11 +119,11 @@ func TestItCanTranscode(t *testing.T) {
 			},
 		},
 		"streamName",
-		clients.InputVideo{
+		video.InputVideo{
 			Duration:  123.0,
 			Format:    "some-format",
 			SizeBytes: 123,
-			Tracks: []clients.InputTrack{
+			Tracks: []video.InputTrack{
 				{
 					Type:       "video",
 					VideoTrack: sourceVideoTrack,
